refactor: 오타 수정 및 트랜잭션 적용
