+++ src/main/java/roomescape/service/booking/reservation/ReservationService.java
@@ -7,27 +7,27 @@ import roomescape.dto.reservation.ReservationRequest;
 import roomescape.dto.reservation.ReservationResponse;
 import roomescape.dto.reservation.UserReservationResponse;
 import roomescape.service.booking.reservation.module.ReservationCancelService;
-import roomescape.service.booking.reservation.module.ReservationResisterService;
+import roomescape.service.booking.reservation.module.ReservationRegisterService;
 import roomescape.service.booking.reservation.module.ReservationSearchService;
 
 @Service
 public class ReservationService {
 
     private final ReservationCancelService reservationCancelService;
-    private final ReservationResisterService reservationResisterService;
+    private final ReservationRegisterService reservationRegisterService;
     private final ReservationSearchService reservationSearchService;
 
     public ReservationService(ReservationCancelService reservationCancelService,
-                              ReservationResisterService reservationResisterService,
+                              ReservationRegisterService reservationRegisterService,
                               ReservationSearchService reservationSearchService
     ) {
         this.reservationCancelService = reservationCancelService;
-        this.reservationResisterService = reservationResisterService;
+        this.reservationRegisterService = reservationRegisterService;
         this.reservationSearchService = reservationSearchService;
     }
 
     public Long resisterReservation(ReservationRequest request) {
-        return reservationResisterService.resisterReservation(request);
+        return reservationRegisterService.registerReservation(request);
     }
 
     public ReservationResponse findReservation(Long reservationId) {
