+++ src/main/java/roomescape/service/booking/reservation/module/ReservationSearchService.java
@@ -2,6 +2,7 @@ package roomescape.service.booking.reservation.module;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.reservation.Reservation;
 import roomescape.domain.waiting.Waiting;
 import roomescape.dto.reservation.ReservationFilter;
@@ -11,6 +12,7 @@ import roomescape.repository.ReservationRepository;
 import roomescape.repository.WaitingRepository;
 
 @Service
+@Transactional(readOnly = true)
 public class ReservationSearchService {
 
     private final ReservationRepository reservationRepository;
@@ -55,7 +57,7 @@ public class ReservationSearchService {
 
     private UserReservationResponse createUserReservationResponse(Reservation reservation) {
         if (reservation.isReserved()) {
-            return UserReservationResponse.creat(reservation);
+            return UserReservationResponse.create(reservation);
         }
 
         Waiting waiting = findWaitingByReservationId(reservation.getId());
