+++ src/main/java/roomescape/service/booking/waiting/WaitingService.java
@@ -2,36 +2,34 @@ package roomescape.service.booking.waiting;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.domain.reservation.Reservation;
-import roomescape.domain.reservation.Status;
 import roomescape.dto.reservation.ReservationRequest;
 import roomescape.dto.waiting.WaitingResponse;
 import roomescape.service.booking.reservation.module.ReservationCancelService;
 import roomescape.service.booking.waiting.module.WaitingCancelService;
-import roomescape.service.booking.waiting.module.WaitingResisterService;
+import roomescape.service.booking.waiting.module.WaitingRegisterService;
 import roomescape.service.booking.waiting.module.WaitingSearchService;
 
 @Service
 public class WaitingService {
 
     private final ReservationCancelService reservationCancelService;
-    private final WaitingResisterService waitingResisterService;
+    private final WaitingRegisterService waitingRegisterService;
     private final WaitingSearchService waitingSearchService;
     private final WaitingCancelService waitingCancelService;
 
     public WaitingService(ReservationCancelService reservationCancelService,
-                          WaitingResisterService waitingResisterService,
+                          WaitingRegisterService waitingRegisterService,
                           WaitingSearchService waitingSearchService,
                           WaitingCancelService waitingCancelService
     ) {
         this.reservationCancelService = reservationCancelService;
-        this.waitingResisterService = waitingResisterService;
+        this.waitingRegisterService = waitingRegisterService;
         this.waitingSearchService = waitingSearchService;
         this.waitingCancelService = waitingCancelService;
     }
 
     public Long resisterWaiting(ReservationRequest request) {
-        return waitingResisterService.registerWaiting(request);
+        return waitingRegisterService.registerWaiting(request);
     }
 
     public List<WaitingResponse> findAllWaitingReservations() {
