+++ src/main/java/roomescape/service/booking/waiting/module/WaitingSearchService.java
@@ -2,10 +2,12 @@ package roomescape.service.booking.waiting.module;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import roomescape.dto.waiting.WaitingResponse;
 import roomescape.repository.WaitingRepository;
 
 @Service
+@Transactional(readOnly = true)
 public class WaitingSearchService {
 
     private final WaitingRepository waitingRepository;
