+++ src/main/java/roomescape/service/member/MemberService.java
@@ -2,11 +2,13 @@ package roomescape.service.member;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.member.Member;
 import roomescape.dto.member.MemberResponse;
 import roomescape.repository.MemberRepository;
 
 @Service
+@Transactional(readOnly = true)
 public class MemberService {
 
     private final MemberRepository memberRepository;
